Assignment 43: first attempt

+++ app/controllers/comments_controller.rb
@@ -5,29 +5,58 @@ before_action :authorize_user, only: [:destroy]
 
 
    def create
-     @post = Post.find(params[:post_id])
-     comment = @post.comments.new(comment_params)
-     comment.user = current_user
+     if @post then
+       @post = Post.find(params[:post_id])
+       comment = @post.comments.new(comment_params)
+       comment.user = current_user
 
-     if comment.save
-       flash[:notice] = "Comment saved successfully."
-       redirect_to [@post.topic, @post]
-     else
-       flash[:alert] = "Comment failed to save."
-       redirect_to [@post.topic, @post]
+       if comment.save
+         flash[:notice] = "Comment saved successfully."
+         redirect_to [@post.topic, @post]
+       else
+         flash[:alert] = "Comment failed to save."
+         redirect_to [@post.topic, @post]
+       end
+     elsif @topic then
+       @topic = Topic.find(params[:topic_id])
+       comment = @topic.comments.new(comment_params)
+       comment.user = current_user
+
+       if comment.save
+         flash[:notice] = "Comment saved successfully."
+         redirect_to [@topic]
+       else
+         flash[:alert] = "Comment failed to save."
+         redirect_to [@topic]
+       end
      end
+
+
    end
 
    def destroy
-      @post = Post.find(params[:post_id])
-      comment = @post.comments.find(params[:id])
+     if @post then
+        @post = Post.find(params[:post_id])
+        comment = @post.comments.find(params[:id])
 
-      if comment.destroy
-        flash[:notice] = "Comment was deleted successfully."
-        redirect_to [@post.topic, @post]
-      else
-        flash[:alert] = "Comment couldn't be deleted. Try again."
-        redirect_to [@post.topic, @post]
+        if comment.destroy
+          flash[:notice] = "Comment was deleted successfully."
+          redirect_to [@post.topic, @post]
+        else
+          flash[:alert] = "Comment couldn't be deleted. Try again."
+          redirect_to [@post.topic, @post]
+        end
+      elsif @topic then
+          @topic = Topic.find(params[:topic_id])
+          comment = @topic.comments.find(params[:id])
+
+           if comment.destroy
+             flash[:notice] = "Comment was deleted successfully."
+             redirect_to [@topic]
+           else
+             flash[:alert] = "Comment couldn't be deleted. Try again."
+             redirect_to [@topic]
+           end
       end
     end
